requirements.txt
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-compat==1.0.15
django-crispy-forms==1.8.0
django-hijack==2.1.10
django-hijack-admin==2.1.10
django-import-export==1.2.0
django-registration==3.0.1
django-storages==1.7.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.4
odfpy==1.4.0
openpyxl==3.0.0
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
